+++ source/DocumentValidation/CimXml/Schemas/urn-ediel-org-measure-notifyaggregatedmeasuredata-0-1.xsd
