+++ source/DocumentValidation/CimXml/Schemas/urn-ediel-org-structure-accountingpointcharacteristics-0-1.xsd
