+++ source/DocumentValidation/CimXml/Schemas/urn-ediel-org-structure-characteristicsofacustomeratanap-0-1.xsd
