+++ source/DocumentValidation/CimXml/Schemas/urn-ediel-org-structure-confirmrequestchangeaccountingpointcharacteristics-0-1.xsd
