+++ source/DocumentValidation/CimXml/Schemas/urn-ediel-org-structure-requestchangecustomercharacteristics-0-1.xsd
