+++ source/DocumentValidation/CimXml/Schemas/urn-entsoe-eu-local-extension-types.xsd
