+++ source/DocumentValidation/CimXml/Schemas/urn-entsoe-eu-wgedi-codelists.xsd
