+++ source/Infrastructure/Configuration/IncomingMessageParsingServices.cs
@@ -19,6 +19,7 @@ using CimMessageAdapter.Messages.RequestChangeCustomerCharacteristics;
 using CimMessageAdapter.Messages.RequestChangeOfSupplier;
 using CimMessageAdapter.Response;
 using DocumentValidation;
+using DocumentValidation.CimXml;
 using Infrastructure.IncomingMessages.RequestChangeOfSupplier;
 using Infrastructure.IncomingMessages.Response;
 using Microsoft.Extensions.DependencyInjection;
