+++ source/Infrastructure/IncomingMessages/RequestChangeOfSupplier/XmlMessageParser.cs
@@ -22,6 +22,7 @@ using Application.IncomingMessages.RequestChangeOfSupplier;
 using CimMessageAdapter.Errors;
 using CimMessageAdapter.Messages;
 using DocumentValidation;
+using DocumentValidation.CimXml;
 using Domain.OutgoingMessages;
 
 namespace Infrastructure.IncomingMessages.RequestChangeOfSupplier;
