+++ source/Tests/Infrastructure/OutgoingMessages/NotifyAggreagtedMeasureData/NotifyAggregatedMeasureDataDocumentWriterTests.cs
@@ -20,7 +20,7 @@ using System.Threading.Tasks;
 using System.Xml.Schema;
 using Application.OutgoingMessages.Common;
 using DocumentValidation;
-using DocumentValidation.Validators;
+using DocumentValidation.CimXml;
 using Domain.Actors;
 using Domain.OutgoingMessages;
 using Domain.OutgoingMessages.NotifyAggregatedMeasureData;
